Move ChemrepoLabels to elements_list folder
