+++ app/packs/src/components/elements_list/ElementsTableSampleEntries.js
@@ -8,7 +8,7 @@ import ElementReactionLabels from 'src/components/elements_list/ElementReactionL
 import ElementWellplateLabels from 'src/components/elements_list/ElementWellplateLabels';
 import GenericElementLabels from 'src/components/generic/GenericElementLabels';
 import PubchemLabels from 'src/components/pubchem/PubchemLabels';
-import ChemrepoLabels from 'src/components/ChemrepoLabels';
+import ChemrepoLabels from 'src/components/elements_list/ChemrepoLabels';
 import ComputedPropLabel from 'src/components/computed_props/ComputedPropLabel';
 import ArrayUtils from 'src/utilities/ArrayUtils';
 import ElementContainer from 'src/components/elements_list/ElementContainer';
